requirements.txt
pyspark==3.3.0
dbt-postgres==1.5.0
great-expectations==0.17.15
pandas==1.5.2
requests==2.28.1
sqlalchemy==1.4.39
psycopg2-binary==2.9.3
tenacity==8.0.1
python-dotenv==0.19.2
rich==12.5.1
typer==0.6.1
jsonschema==4.4.0
cryptography>=41.0.0
